TROLL 4.0
This new version of the TROLL model integrates water fluxes, vegetation responses to water availability, a refined simulation of leaf phenology and carbon allocation, intraspecific trait variation as well as more realistic tree crowns. It is described in details and evaluated in two companions manuscripts submitted to Geoscientific Model Development (Discussion) in october 2024.